requirements.txt
Pillow
django-filebrowser-no-grappelli
django-tinymce
